ssize_t is 32-bit on ARM/32 ... cast to suppress warning:
src/net.c: In function 'icmp_echo':
src/net.c:447:33: warning: format '%ld' expects argument of type 'long int', but argument 5 has type 'ssize_t' [-Wformat]

+++ src/net.c
@@ -444,7 +444,7 @@ readnext:
                                 LogError("Ping response for %s %d/%d failed -- %s\n", hostname, i + 1, count, STRERROR);
                                 continue;
                         } else if (n < in_len) {
-                                LogError("Ping response for %s %d/%d failed -- received %ld bytes, expected at least %d bytes\n", hostname, i + 1, count, n, in_len);
+                                LogError("Ping response for %s %d/%d failed -- received %ld bytes, expected at least %d bytes\n", hostname, i + 1, count, (long)n, in_len);
                                 continue;
                         }
                         boolean_t in_addrmatch = false, in_typematch = false;
